+++ evil_flowers_catalog/settings/base.py
@@ -194,7 +194,7 @@ STATIC_URL = "/static/"
 STATIC_ROOT = os.path.join(BASE_DIR, "static")
 
 MEDIA_URL = "/media/"
-MEDIA_ROOT = os.path.join(BASE_DIR, "media")
+MEDIA_ROOT = os.getenv("MEDIA_ROOT", os.path.join(BASE_DIR, "media"))
 
 DATA_UPLOAD_MAX_MEMORY_SIZE = 1024 * 1024 * 100  # 100MB