+++ CHANGELOG.md
@@ -5,7 +5,9 @@ This changelog suppose to follow rules defined in the [changelog.md](https://cha
 ## 0.9.0 : TBD
 
 - **Added**: Application cache layer
+- **Added**: `/data/v1/thumbnails/:entry_id` introduced
 - **Changed**: Sort entries in Complete OPDS catalog by `created_at`
+- **Changed**: `EntrySerializer` now returns URL path in the `thumbnail` property instead of base64 image (performance)
 
 ## 0.8.0 : 2024-05-09